+++ examples/start-electron-app/src/electron/preload/main.preload.ts
@@ -1,3 +1,2 @@
 import { contextBridge, ipcRenderer } from 'electron';
-
 contextBridge.exposeInMainWorld('ipc', ipcRenderer);