+++ packages/framework/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @electron-boot/framework
 
+## 1.0.0-beta.7
+
+### Major Changes
+
+- fix: get request ctx
+
+### Patch Changes
+
+- Updated dependencies
+  - @electron-boot/logger@1.0.0-beta.7
+
+## 1.0.0-beta.6
+
+### Minor Changes
+
+- perf: Streamline Code
+
+### Patch Changes
+
+- Updated dependencies
+  - @electron-boot/logger@1.0.0-beta.6
+
 ## 1.0.0-beta.5
 
 ### Major Changes