+++ packages/logger/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @electron-boot/logger
 
+## 1.0.0-beta.7
+
+### Major Changes
+
+- fix: get request ctx
+
+## 1.0.0-beta.6
+
+### Minor Changes
+
+- perf: Streamline Code
+
 ## 1.0.0-beta.5
 
 ### Major Changes